+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Make sure disposing a node removes it from the SCM and it's scope
 - SubScopeManager can dynamically create new sub scopes
 - Node and scope names must be camel case now
+- Scope:instantiate: Rename parentScope param into scope
 
 ### Removed